requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
SQLAlchemy==2.0.23
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
imagehash==4.3.1
scikit-learn==1.3.2

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
responses==0.24.1

# Logging enhancements (optional)
colorama==0.4.6
